+++ src/collective/fhirpath/tests/test_fhirpath_engine.py
@@ -32,7 +32,7 @@ class FhirpathPloneEngineIntegrationTest(unittest.TestCase):
         factory = queryMultiAdapter(
             (self.get_es_catalog(),), IElasticsearchEngineFactory
         )
-        engine = factory(fhir_version=FHIR_VERSION.STU3)
+        engine = factory(fhir_release=FHIR_VERSION.STU3)
         return engine
 
     def test_calculate_field_index_name(self):