+++ src/collective/fhirpath/tests/test_fhirpath_query.py
@@ -31,7 +31,7 @@ class FhirPathPloneQueryFunctionalTest(BaseFunctionalTesting):
         factory = queryMultiAdapter(
             (self.get_es_catalog(),), IElasticsearchEngineFactory
         )
-        engine = factory(fhir_version=FHIR_VERSION.STU3)
+        engine = factory(fhir_release=FHIR_VERSION.STU3)
         return engine
 
     def test_iter_result(self):