+++ src/collective/fhirpath/tests/test_pluginindexes.py
@@ -1,7 +1,6 @@
 # _*_ coding: utf-8 _*_
 from collective.fhirpath import utils
 from collective.fhirpath.testing import COLLECTIVE_FHIRPATH_INTEGRATION_TESTING
-from fhirpath.enums import FHIR_VERSION
 
 import unittest
 
@@ -23,6 +22,6 @@ class PluginIndexesIntegrationTest(unittest.TestCase):
         catalogtool = getattr(self.portal, "portal_catalog")
         catalogtool._catalog.getIndex("organization_resource")
         mapping1 = catalogtool._catalog.getIndex("organization_resource").mapping
-        mapping2 = utils.get_elasticsearch_mapping("Organization", FHIR_VERSION.STU3)
+        mapping2 = utils.get_elasticsearch_mapping("Organization", "STU3")
 
         self.assertEqual(mapping1, mapping2)